d3m_ibex/requirements.txt
setuptools
ipykernel
jupyter
matplotlib
numpy
scipy>=0.19.0
sklearn
pandas>=0.19.2
dill
kafka
flask
requests
nltk
grpcio-tools
flask
click # must install manually - `pip3 install click`, won't work through requirements.txt file 
spacy==2.1.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.1.0/en_core_web_md-2.1.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/es_core_news_md-2.1.0/es_core_news_md-2.1.0.tar.gz


